split Alterable to LogicAndActualTablesAware and RoutingUnitAware
