+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/token/generator/impl/DistinctProjectionPrefixTokenGenerator.java
@@ -21,7 +21,7 @@ import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementConte
 import org.apache.shardingsphere.sql.parser.relation.statement.impl.SelectSQLStatementContext;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.token.generator.IgnoreForSingleRoute;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.OptionalSQLTokenGenerator;
-import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.DistinctProjectionPrefixToken;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.impl.DistinctProjectionPrefixToken;
 
 /**
  * Distinct projection prefix token generator.
