+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/token/generator/impl/keygen/GeneratedKeyInsertColumnTokenGenerator.java
@@ -22,7 +22,7 @@ import com.google.common.base.Preconditions;
 import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementContext;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.column.InsertColumnsSegment;
 import org.apache.shardingsphere.sql.parser.sql.statement.dml.InsertStatement;
-import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.GeneratedKeyInsertColumnToken;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.impl.GeneratedKeyInsertColumnToken;
 import org.apache.shardingsphere.core.route.router.sharding.keygen.GeneratedKey;
 
 /**
