+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/token/pojo/impl/LiteralGeneratedKeyAssignmentToken.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.impl;
+
+/**
+ * Generated key assignment token for literal.
+ *
+ * @author panjuan
+ */
+public final class LiteralGeneratedKeyAssignmentToken extends GeneratedKeyAssignmentToken {
+    
+    private final Object value;
+    
+    public LiteralGeneratedKeyAssignmentToken(final int startIndex, final String columnName, final Object value) {
+        super(startIndex, columnName);
+        this.value = value;
+    }
+    
+    @Override
+    protected String getRightValue() {
+        return value instanceof String ? String.format("'%s'", value) : value.toString();
+    }
+}
