+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/token/pojo/impl/TableToken.java
@@ -15,16 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo;
+package org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.impl;
 
 import com.google.common.base.Joiner;
 import lombok.Getter;
-import org.apache.shardingsphere.sql.parser.core.constant.QuoteCharacter;
-import org.apache.shardingsphere.sql.parser.util.SQLUtil;
-import org.apache.shardingsphere.core.rewrite.sql.token.pojo.Alterable;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.token.pojo.LogicAndActualTablesAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.SQLToken;
 import org.apache.shardingsphere.core.rewrite.sql.token.pojo.Substitutable;
-import org.apache.shardingsphere.core.route.type.RoutingUnit;
+import org.apache.shardingsphere.sql.parser.core.constant.QuoteCharacter;
+import org.apache.shardingsphere.sql.parser.util.SQLUtil;
 
 import java.util.Map;
 
@@ -34,7 +33,7 @@ import java.util.Map;
  * @author zhangliang
  * @author panjuan
  */
-public final class TableToken extends SQLToken implements Substitutable, Alterable {
+public final class TableToken extends SQLToken implements Substitutable, LogicAndActualTablesAware {
     
     @Getter
     private final int stopIndex;
@@ -56,7 +55,7 @@ public final class TableToken extends SQLToken implements Substitutable, Alterab
     }
     
     @Override
-    public String toString(final RoutingUnit routingUnit, final Map<String, String> logicAndActualTables) {
+    public String toString(final Map<String, String> logicAndActualTables) {
         String actualTableName = logicAndActualTables.get(tableName.toLowerCase());
         actualTableName = null == actualTableName ? tableName.toLowerCase() : actualTableName;
         return Joiner.on("").join(quoteCharacter.getStartDelimiter(), actualTableName, quoteCharacter.getEndDelimiter());
